feat(#2): add mode setter and getter to interface
